Added new mode, details in description
Update executable.

Added a new game mode:
now the goal is to destroy the eggs before they hatch, all the alien born will act like the mother (the mother is not present), after the hatch you will have to survive 1 minute till the rescue arrive.

Implemented a loading image due to an loading freeze happening when the game scene appears without understanding where to handle it with a ResourceLoader with progress bar.

Implemented a progress bar to the skip video button